requirements.txt
python-telegram-bot
google-api-python-client
wordcloud
